Separate server related settings into another class
