+++ app/src/main/java/dev/bartuzen/qbitcontroller/ui/addtorrent/AddTorrentViewModel.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dagger.hilt.android.qualifiers.ApplicationContext
-import dev.bartuzen.qbitcontroller.data.SettingsManager
+import dev.bartuzen.qbitcontroller.data.ServerManager
 import dev.bartuzen.qbitcontroller.data.repositories.AddTorrentRepository
 import dev.bartuzen.qbitcontroller.model.ServerConfig
 import dev.bartuzen.qbitcontroller.network.RequestError
@@ -28,7 +28,7 @@ import javax.inject.Inject
 class AddTorrentViewModel @Inject constructor(
     @ApplicationContext private val context: Context,
     private val repository: AddTorrentRepository,
-    private val settingsManager: SettingsManager
+    private val serverManager: ServerManager
 ) : ViewModel() {
     private val eventChannel = Channel<Event>()
     val eventFlow = eventChannel.receiveAsFlow()
@@ -50,7 +50,7 @@ class AddTorrentViewModel @Inject constructor(
 
     var isInitialLoadStarted = false
 
-    fun getServers() = settingsManager.serversFlow.value.values.toList()
+    fun getServers() = serverManager.serversFlow.value.values.toList()
 
     fun createTorrent(
         serverConfig: ServerConfig,
