+++ app/src/main/java/dev/bartuzen/qbitcontroller/ui/main/MainActivity.kt
@@ -109,7 +109,7 @@ class MainActivity : AppCompatActivity() {
             startActivity(intent)
         }
 
-        viewModel.serverList.launchAndCollectLatestIn(this) { serverList ->
+        viewModel.serversFlow.launchAndCollectLatestIn(this) { serverList ->
             serverListAdapter.submitList(serverList.values.toList())
 
             binding.textClickToAddServer.visibility =
