+++ app/src/main/java/dev/bartuzen/qbitcontroller/ui/main/MainViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import dev.bartuzen.qbitcontroller.data.SettingsManager
+import dev.bartuzen.qbitcontroller.data.ServerManager
 import dev.bartuzen.qbitcontroller.model.ServerConfig
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
@@ -13,11 +13,11 @@ import javax.inject.Inject
 @HiltViewModel
 class MainViewModel @Inject constructor(
     private val state: SavedStateHandle,
-    settingsManager: SettingsManager
+    serverManager: ServerManager
 ) : ViewModel() {
     val currentServer = state.getStateFlow<ServerConfig?>("current_server", null)
 
-    val serverList = settingsManager.serversFlow
+    val serversFlow = serverManager.serversFlow
 
     fun setCurrentServer(serverConfig: ServerConfig) {
         state["current_server"] = serverConfig
@@ -25,7 +25,7 @@ class MainViewModel @Inject constructor(
 
     init {
         viewModelScope.launch {
-            serverList.collectLatest { serverList ->
+            serversFlow.collectLatest { serverList ->
                 val currentServerId = currentServer.value?.id ?: -1
                 val firstServer = try {
                     serverList[serverList.firstKey()]
