+++ app/src/main/java/dev/bartuzen/qbitcontroller/ui/settings/SettingsViewModel.kt
@@ -2,12 +2,12 @@ package dev.bartuzen.qbitcontroller.ui.settings
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
-import dev.bartuzen.qbitcontroller.data.SettingsManager
+import dev.bartuzen.qbitcontroller.data.ServerManager
 import javax.inject.Inject
 
 @HiltViewModel
 class SettingsViewModel @Inject constructor(
-    private val settingsManager: SettingsManager
+    private val serverManager: ServerManager
 ) : ViewModel() {
-    fun getServers() = settingsManager.serversFlow.value
+    fun getServers() = serverManager.serversFlow.value
 }
