+++ app/src/main/java/dev/bartuzen/qbitcontroller/ui/settings/addeditserver/AddEditServerViewModel.kt
@@ -3,7 +3,7 @@ package dev.bartuzen.qbitcontroller.ui.settings.addeditserver
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import dev.bartuzen.qbitcontroller.data.SettingsManager
+import dev.bartuzen.qbitcontroller.data.ServerManager
 import dev.bartuzen.qbitcontroller.model.Protocol
 import dev.bartuzen.qbitcontroller.model.ServerConfig
 import kotlinx.coroutines.launch
@@ -11,7 +11,7 @@ import javax.inject.Inject
 
 @HiltViewModel
 class AddEditServerViewModel @Inject constructor(
-    private val settingsManager: SettingsManager
+    private val serverManager: ServerManager
 ) : ViewModel() {
     fun addServer(
         name: String?,
@@ -22,14 +22,14 @@ class AddEditServerViewModel @Inject constructor(
         username: String,
         password: String
     ) = viewModelScope.launch {
-        settingsManager.addServer(name, protocol, host, port, path, username, password)
+        serverManager.addServer(name, protocol, host, port, path, username, password)
     }
 
     fun editServer(serverConfig: ServerConfig) = viewModelScope.launch {
-        settingsManager.editServer(serverConfig)
+        serverManager.editServer(serverConfig)
     }
 
     fun removeServer(serverConfig: ServerConfig) = viewModelScope.launch {
-        settingsManager.removeServer(serverConfig)
+        serverManager.removeServer(serverConfig)
     }
 }
